Fix `/autolink test` command for disabled links (#123)
Also:
- Started a "library" of domain-specific tests, to be used as templates for configuration
- Added a template for ProductBoard
- Fixed a lint nit in client.go
